Aesthetic refactoring of Token
Initially, there was an EOF type for which a position didn't really make
sense since it always had to be at the end.
So, all other token matches had a position associated with them.
With the change to EOF being signalled by exhaustion of the tokenizing
iterator instead, all Token branches had a pos.

This changes Token to be a struct containing a Data enum (the previous
token type) and a pos